requirements.txt
chainlit==1.0.506
openai==1.74.0
python-dotenv==1.0.1
